Separate user registration view functions into their own module
